+++ Library/Dsqt/model/dsReferenceMap.h
