+++ Library/Dsqt/qml/waffles/media_viewer/DsControlSet.qml
